+++ source/NFmiPressManager.cpp
@@ -243,7 +243,7 @@ bool NFmiPressManager::ReadDescriptionAndWrite(NFmiPressProduct & thePressProduc
 			  *itsLogFile << "*** ERROR: Not valid PressMask : "
 						  << static_cast<char *>(maskName)
 						  << " in space "
-						  << helpEnumSpace 
+						  << static_cast<int>(helpEnumSpace)
 						  << endl;
 			}
